refactor(analysis): split scope helpers
