+++ internal/analysis/scope.go
@@ -1,15 +1,11 @@
 package analysis
 
 import (
-	"errors"
 	"fmt"
-	"io"
 	"io/fs"
 	"os"
 	"path/filepath"
 	"regexp"
-	"sort"
-	"strings"
 )
 
 const maxScopeDiagnostics = 5
@@ -165,166 +161,4 @@ func recordScopeSkipReason(stats *scopeStats, slashed string, reason string) {
 		return
 	}
 	stats.skippedDiagnostics = append(stats.skippedDiagnostics, slashed+" ("+reason+")")
-}
-
-func normalizePatterns(patterns []string) []string {
-	seen := make(map[string]struct{}, len(patterns))
-	result := make([]string, 0, len(patterns))
-	for _, pattern := range patterns {
-		trimmed := strings.TrimSpace(pattern)
-		if trimmed == "" {
-			continue
-		}
-		if _, exists := seen[trimmed]; exists {
-			continue
-		}
-		seen[trimmed] = struct{}{}
-		result = append(result, filepath.ToSlash(trimmed))
-	}
-	return result
-}
-
-func compileGlobPatterns(patterns []string) ([]compiledPattern, error) {
-	compiled := make([]compiledPattern, 0, len(patterns))
-	for _, pattern := range patterns {
-		regex, err := regexp.Compile(globToRegexp(pattern))
-		if err != nil {
-			return nil, fmt.Errorf("compile scope pattern %q: %w", pattern, err)
-		}
-		compiled = append(compiled, compiledPattern{
-			pattern: pattern,
-			regex:   regex,
-		})
-	}
-	return compiled, nil
-}
-
-func matchFirstCompiledPattern(path string, patterns []compiledPattern) (bool, string) {
-	for _, pattern := range patterns {
-		if pattern.regex.MatchString(path) {
-			return true, pattern.pattern
-		}
-	}
-	return false, ""
-}
-
-func globMatch(pattern, value string) bool {
-	expression := globToRegexp(pattern)
-	matched, err := regexp.MatchString(expression, value)
-	return err == nil && matched
-}
-
-func globToRegexp(pattern string) string {
-	var builder strings.Builder
-	builder.Grow(len(pattern) * 2)
-	builder.WriteString("^")
-	for index := 0; index < len(pattern); index++ {
-		char := pattern[index]
-		if char == '*' {
-			segment, next := asteriskSegment(pattern, index)
-			builder.WriteString(segment)
-			index = next
-			continue
-		}
-		if char == '?' {
-			builder.WriteString("[^/]")
-			continue
-		}
-		if strings.ContainsRune(`.+()|[]{}^$\\`, rune(char)) {
-			builder.WriteByte('\\')
-		}
-		builder.WriteByte(char)
-	}
-	builder.WriteString("$")
-	return builder.String()
-}
-
-func asteriskSegment(pattern string, index int) (string, int) {
-	if index+1 < len(pattern) && pattern[index+1] == '*' {
-		if index+2 < len(pattern) && pattern[index+2] == '/' {
-			return "(?:.*/)?", index + 2
-		}
-		return ".*", index + 1
-	}
-	return "[^/]*", index
-}
-
-func formatPatternMatches(patterns []string, matches map[string]int) string {
-	parts := make([]string, 0, len(patterns))
-	for _, pattern := range patterns {
-		parts = append(parts, fmt.Sprintf("%s=%d", pattern, matches[pattern]))
-	}
-	sort.Strings(parts)
-	return strings.Join(parts, ", ")
-}
-
-func copyFile(repoPath, scopedRoot, relativePath string) (err error) {
-	if !isSafeRelativePath(relativePath) {
-		return fmt.Errorf("invalid relative path for scoped copy: %s", relativePath)
-	}
-	cleanRelativePath := filepath.Clean(relativePath)
-	sourcePath := filepath.Join(repoPath, cleanRelativePath)
-	targetPath := filepath.Join(scopedRoot, cleanRelativePath)
-	if !pathWithin(repoPath, sourcePath) {
-		return fmt.Errorf("source path escapes repository scope: %s", sourcePath)
-	}
-	if !pathWithin(scopedRoot, targetPath) {
-		return fmt.Errorf("target path escapes scoped workspace: %s", targetPath)
-	}
-	if err := os.MkdirAll(filepath.Dir(targetPath), 0o750); err != nil {
-		return err
-	}
-
-	sourceRoot, err := os.OpenRoot(repoPath)
-	if err != nil {
-		return fmt.Errorf("open source root: %w", err)
-	}
-	defer joinCloseError(&err, sourceRoot.Close)
-
-	source, err := sourceRoot.Open(cleanRelativePath)
-	if err != nil {
-		return err
-	}
-	defer joinCloseError(&err, source.Close)
-
-	targetRoot, err := os.OpenRoot(scopedRoot)
-	if err != nil {
-		return fmt.Errorf("open target root: %w", err)
-	}
-	defer joinCloseError(&err, targetRoot.Close)
-
-	target, err := targetRoot.OpenFile(cleanRelativePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o600)
-	if err != nil {
-		return err
-	}
-	defer joinCloseError(&err, target.Close)
-	if _, err := io.Copy(target, source); err != nil {
-		return err
-	}
-	return nil
-}
-
-func joinCloseError(target *error, closeFn func() error) {
-	if closeErr := closeFn(); closeErr != nil {
-		*target = errors.Join(*target, closeErr)
-	}
-}
-
-func pathWithin(root, candidate string) bool {
-	relative, err := filepath.Rel(root, candidate)
-	if err != nil {
-		return false
-	}
-	return relative == "." || (relative != ".." && !strings.HasPrefix(relative, ".."+string(filepath.Separator)))
-}
-
-func isSafeRelativePath(relativePath string) bool {
-	if filepath.IsAbs(relativePath) {
-		return false
-	}
-	cleaned := filepath.Clean(relativePath)
-	if cleaned == "." {
-		return false
-	}
-	return cleaned != ".." && !strings.HasPrefix(cleaned, ".."+string(filepath.Separator))
 }
